panda_ai/panda_ai/requirements.txt
flask
flask-cors
python-dotenv
requests
feedparser
trafilatura
gunicorn
pdfplumber
python-docx
openpyxl
python-pptx
pytesseract
Pillow
openai-whisper
moviepy
SpeechRecognition